Updated Entity and fixed some bugs

+++ backend/src/main/java/com/bankrupted/greenroof/service/forum/ForumAnswerService.java
@@ -58,21 +58,24 @@ public class ForumAnswerService {
         return new ResponseEntity<>("Updated Answer", HttpStatus.CREATED);
     }
 
-    public ResponseEntity<?> deleteAnswerOfQuestion(Long questionId) {
-        ForumAnswer forumAnswer = forumAnswerRepository.findTop1ByQuestionId(questionId).get();
-        Long answerId = forumAnswer.getId();
+    public void deleteAnswerOfQuestion(Long questionId) {
+        Boolean doesExist = forumAnswerRepository.existsByQuestionId(questionId);
+        if(!doesExist)
+            return;
+
+        Long answerId = forumAnswerRepository.findByQuestionIdOrderByScoreDescCreatedAtDesc(questionId).get(0).getId();
         Integer voteListSize = forumVoteRepository.findByAnswerId(answerId).size();
         if(voteListSize > 0) forumVoteRepository.deleteByAnswerId(answerId);
         forumAnswerRepository.deleteByQuestionId(questionId);
-        return new ResponseEntity<>("Answer deleted successfully.", HttpStatus.OK);
     }
 
     public ResponseEntity<?> deleteAnswerOfQuestion(String username, Long answerId) {
         ForumAnswer forumAnswer = forumAnswerRepository.findById(answerId)
                         .orElseThrow(() -> new NoSuchElementException("Answer with id " + answerId + " does not exists."));
-        System.out.println(forumAnswer);
+
         if(!Objects.equals(forumAnswer.getAnswerer().getUsername(), username))
             return new ResponseEntity<>("You are not allowed to delete this answer.", HttpStatus.FORBIDDEN);
+
         if(!forumVoteRepository.findByAnswerId(answerId).isEmpty()) forumVoteRepository.deleteByAnswerId(answerId);
         forumAnswerRepository.deleteById(answerId);
         return new ResponseEntity<>("Answer deleted successfully.", HttpStatus.OK);
@@ -81,6 +84,7 @@ public class ForumAnswerService {
     public ResponseEntity<?> getAnswersOfSingleQuestion(Long questionId) {
         forumQuestionRepository.findById(questionId)
                 .orElseThrow(() -> new NoSuchElementException("Question with id " + questionId + " does not exists."));
+
         List<ForumAnswer> answers = forumAnswerRepository.findByQuestionIdOrderByScoreDescCreatedAtDesc(questionId);
         answers.forEach(answer -> {
             answer.setScore(forumVoteRepository.getTotalVotesOfAnswer(answer.getId()));
